+++ Stop-OpenSshServer.ps1
@@ -1,3 +1,14 @@
+<#PSScriptInfo
+.VERSION 0.2.1
+.GUID d6e03cb4-a92f-4550-bdda-81093864c6a4
+.AUTHOR metyatech
+.COMPANYNAME metyatech
+.COPYRIGHT (c) 2026 metyatech. All rights reserved.
+.TAGS OpenSSH, SSH, Windows, Server, Stop
+.LICENSEURI https://github.com/metyatech/windows-openssh-server-startup/blob/main/LICENSE
+.PROJECTURI https://github.com/metyatech/windows-openssh-server-startup
+.RELEASENOTES Add PowerShell Gallery metadata for distribution.
+#>
 #requires -Version 5.1
 [CmdletBinding(SupportsShouldProcess = $true, ConfirmImpact = 'Medium')]
 param(